refactor: Split JSON operations.ex into smaller modules by category
Split the large operations.ex file (931 lines) into specialized sub-modules to improve maintainability and navigation:

- PtcRunner.Json.Operations.Comparison: eq, neq, gt, gte, lt, lte, contains (~60 lines)
- PtcRunner.Json.Operations.Aggregation: sum, count, avg, min, max (~160 lines)
- PtcRunner.Json.Operations.Collection: filter, map, select, reject (~120 lines)
- PtcRunner.Json.Operations.Access: get, first, last, nth, sort_by, max_by, min_by (~180 lines)

The main Operations module now acts as a dispatcher, maintaining the existing public API (Operations.eval/4). All tests pass without modification. Implementation mirrors the existing test structure in test/ptc_runner/json/operations/.
